[FEATURE] Allow passing of request options and own error handling
With this change, it is now possible to set options to be passed
with the request method on a per callee level. Each client creation
is able to accept options, that will be applied to each call to
any request function and be respected by Guzzle Requests.
The default behaviour will be as before that change.

Additionally, the exception handling and the Response wrapping
into a SevenShores\Hubspot\Http\Response object can be avoided.
So it is possible to care for the error handling in the own
application.
Again, the default behaviour remains unchanged, the response
is wrapped and SevenShores\Hubspot\Exceptions are thrown.