API fix for folders as well as Guzzle warnings for file replace (#133)
* Fix folders

The folders API call should be in this format. The way it was the options where never being used.

* Guzzle body for file is deprecated

MItigate the following error from Guzzle.

Passing in the "body" request option as an array to send a POST request has been deprecated.

Please use the "form_params" request option to send a application/x-www-form-urlencoded request, or the "multipart" request option to send a multipart/form-data request.