Add support for removing list contacts by email address
The endpoint to remove contacts from a list will also support an array of emails being passed through. This changes the removeContact signature to match the addContact signature by adding an optional 3rd parameter to the removeContact method so that users may be removed from a list by email address.